requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
psycopg2==2.8.3
python-dotenv==0.10.3
python-telegram-bot==12.0.0
requests==2.22.0
SQLAlchemy==1.3.8